% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agencies_v2}
\alias{agencies_v2}
\title{Washington Post Police Agencies Dataset for Shootings Dataset Version 2}
\format{
A data frame with 3103 rows and 6 variables:
\describe{
  \item{\code{id}}{double Department database id}
  \item{\code{name}}{character Department name}
  \item{\code{type}}{character Department type}
  \item{\code{state}}{character Department state}
  \item{\code{oricodes}}{character Department ORI codes (federal identifiers). Note that there can be multiple ORI codes for a single department due to aggregated sub-departments and/or differing values provided by the FBI vs DOJ.}
  \item{\code{total_shootings}}{double Total death records the agency has been involved with.}
}
}
\source{
\url{https://github.com/washingtonpost/data-police-shootings/}
}
\usage{
agencies_v2
}
\description{
Data for departments and agencies involved in a fatal police shooting in the database. This separate CSV can be joined with the death record CSV via the agency_ids value.
}
\keyword{datasets}
