% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapoposh-package.R
\docType{package}
\name{wapoposh-package}
\alias{wapoposh}
\alias{wapoposh-package}
\title{wapoposh: Washington Post Police Shooting Dataset}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides Washington Post Police Shootings dataset. Data source: Washington Post Github \url{https://github.com/washingtonpost/data-police-shootings}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/parmsam/wapoposh}
  \item \url{https://parmsam.github.io/wapoposh/}
  \item Report bugs at \url{https://github.com/parmsam/wapoposh/issues}
}

}
\keyword{internal}
